Isolate OutOfDate UI logic
1. Rename class: EngineVersionProperties->OutOfDateDialog.java
2. Replace property values in OutOfDateDialog.java with equivalent UriConstants.
   Turns out most of the data from latest_engine.properties we already have
   hardcoded as constants, everything except for the 'latest version field'
3. Push 'latest version field' up to caller, EngineVersionCheck
