+++ game-core/src/main/java/games/strategy/engine/auto/update/EngineVersionCheck.java
@@ -5,18 +5,25 @@ import com.google.common.base.Splitter;
 import games.strategy.engine.ClientContext;
 import games.strategy.triplea.settings.ClientSetting;
 import games.strategy.triplea.settings.GameSetting;
+import java.io.IOException;
+import java.net.URL;
 import java.time.LocalDate;
 import java.time.ZoneId;
 import java.util.List;
+import java.util.Properties;
 import java.util.logging.Level;
 import javax.swing.JOptionPane;
 import javax.swing.SwingUtilities;
+import lombok.experimental.UtilityClass;
 import lombok.extern.java.Log;
 import org.triplea.swing.EventThreadJOptionPane;
+import org.triplea.util.Version;
 
 @Log
+@UtilityClass
 final class EngineVersionCheck {
-  private EngineVersionCheck() {}
+  private static final String TRIPLEA_VERSION_LINK =
+      "https://raw.githubusercontent.com/triplea-game/triplea/master/latest_version.properties";
 
   static void checkForLatestEngineVersionOut() {
     try {
@@ -24,20 +31,32 @@ final class EngineVersionCheck {
         return;
       }
 
-      final EngineVersionProperties latestEngineOut = new EngineVersionProperties();
+      final Version latestVersionOut =
+          new Version(
+              getProperties().getProperty("LATEST", ClientContext.engineVersion().toString()));
 
-      if (latestEngineOut.getLatestVersionOut().isGreaterThan(ClientContext.engineVersion())) {
+      if (latestVersionOut.isGreaterThan(ClientContext.engineVersion())) {
         SwingUtilities.invokeLater(
             () ->
                 EventThreadJOptionPane.showMessageDialog(
                     null,
-                    latestEngineOut.getOutOfDateComponent(),
+                    OutOfDateDialog.getOutOfDateComponent(latestVersionOut),
                     "Please Update TripleA",
                     JOptionPane.INFORMATION_MESSAGE));
       }
     } catch (final Exception e) {
       log.log(Level.SEVERE, "Error while checking for engine updates", e);
     }
+  }
+
+  private static Properties getProperties() {
+    final Properties props = new Properties();
+    try {
+      props.load(new URL(TRIPLEA_VERSION_LINK).openStream());
+    } catch (final IOException e) {
+      log.info("Failed to get TripleA latest version file, check internet connection, error: " + e);
+    }
+    return props;
   }
 
   private static boolean isEngineUpdateCheckRequired() {
